backend/requirements.txt
acres==0.5.0
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
ci-info==0.3.0
click==8.2.1
configobj==5.0.9
configparser==7.2.0
distro==1.9.0
etelemetry==0.3.1
filelock==3.19.1
fitz==0.0.1.dev2
Flask==3.1.2
flask-cors==6.0.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
looseversion==1.3.0
lxml==6.0.1
MarkupSafe==3.0.2
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.3
openai==1.107.3
packaging==25.0
pandas==2.3.2
pathlib==1.0.1
pdf2image==1.17.0
pillow==11.3.0
prov==2.1.1
puremagic==1.30
pydantic==2.11.9
pydantic_core==2.33.2
pydot==4.0.1
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyxnat==1.6.3
rdflib==7.1.4
requests==2.32.5
scipy==1.16.2
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
traits==7.0.2
types-Flask-Cors==6.0.0.20250809
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
